Add exeptions workflow

+++ c8/data/explorer.py
@@ -1,6 +1,6 @@
-from .init import curs
+from .init import (curs, conn, IntegrityError)
 from model.explorer import Explorer
-
+from errors import Missing, Duplicate
 
 curs.execute("""create table if not exists explorer(
              name text primary key,
@@ -17,34 +17,52 @@ def get_one(name: str) -> Explorer:
     qry = "select * from explorer where name=:name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 def get_all() -> list[Explorer]:
     qry = "select * from explorer"
     curs.execute(qry)
     return [row_to_model(row) for row in curs.fetchall()]
 
-def create(explorer: Explorer) -> Explorer:
-    qry = """insert into explorer (name, country, description)
-          values (:name, :country, :description)"""
+def create(explorer: Explorer) -> Explorer|None:
+    if not explorer: return None
+    qry = """insert into explorer (name, country, description) values
+             (:name, :country, :description)"""
     params = model_to_dict(explorer)
-    _ = curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+        conn.commit()
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {explorer.name} already exists")
     return get_one(explorer.name)
 
-def modify(name: str, explorer: Explorer) -> Explorer:
+
+def modify(name: str, explorer: Explorer) -> Explorer|None:
+    if not (name and explorer): return None
     qry = """update explorer
-          set country=:country,
-          name=:name,
-          description=:description
-          where name=:name_orig"""
+             set country=:country,
+             name=:name,
+             description=:description
+             where name=:name_orig"""
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    curs.execute(qry, params)
+    conn.commit()
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name: return False
     qry = "delete from explorer where name = :name"
-    params = {"name": explorer.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    conn.commit()
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Explorer {name} not found")
+
